+++ tests/testthat/test-languagetool.R
@@ -31,22 +31,3 @@ load(system.file(package = "LanguageToolR", "expected_output", "expected_output.
 test_that("languagetool JSON parsing works`", {
   expect_true(comparison_func(output_text, output_stored))
 })
-
-test_that("languagetool fails where needed`", {
-  # No input
-  expect_error(languagetool(executable = current_executable), info = "Error in languagetool() : No input defined.")
-  # The 'executable' command is incorrect.
-  if (.Platform$OS.type != "windows") {
-    expect_error(
-      languagetool(test_text[1], tagger_only = TRUE, executable = "-")
-    )
-  } else {
-    # Both warning (by `system()`) and error (by `languagetool()`) are returned
-    # on Windows
-    expect_warning(
-      expect_error(
-        languagetool(test_text[1], tagger_only = TRUE, executable = "-")
-      )
-    )
-  }
-})